Automate the commit checks and move the checklist to CONTRIBUTING.md
We can avoid the dependency on `13rac1/block-fixup-merge-action` because
there is a community-contributed `gitlint` rule for it [0].
The `contrib-disallow-cleanup-commits` rule enforces that the
subject/title does not contain `fixup!`, `squash!`, or `amend!`.
We can also use the `contrib-body-requires-signed-off-by` rule to check
that the body contains the `Signed-off-by` line.

We decided to move the checklist from the PR template to CONTRIBUTING.md
so that contributors don't have to manually tick all checkbox entries
for every PR.
The CI workflow already runs `cargo check`,
`cargo test --all --all-features`, and
`cargo clippy --all --all-targets --all-features -- -D warnings`.

The `github.event_name == 'pull_request'` check isn't required because
of the `on: [pull_request]` statement at the beginning of the file.

[0]: https://jorisroovers.com/gitlint/contrib_rules/

Signed-off-by: Michael Weiss <michael.weiss@atos.net>