actions: Run optional clippy (linting) checks with newer versions
More recent toolchain versions come with additional clippy/linting
checks. E.g., the current stable version is 1.65.0 and shows the three additional
clippy warnings (compared to 1.64.0 / the MSRV). We already saw those
warnings locally (until aa10870).

We only make the check with the MSRV required as contributors shouldn't
have to fix clippy warnings in other parts of the code (we'll take care
of any additional warnings when bumping the MSRV).
However, it is useful to also run the clippy check with a more recent
toolchain version so that we can noticed and fix new warnings in
advance. Most users will use the stable version so using the beta
version for the check seems ideal to catch warnings in advance (so that
users don't see them when they run clippy locally).

Signed-off-by: Michael Weiss <michael.weiss@atos.net>