Update all dependencies
This is just the result of running `cargo update` without touching
Cargo.toml. It'll interfere with dependabot but I think it's a good idea
as this hasn't been done for quite a while and should hopefully make it
easier to manually update some dependencies that dependabot cannot
handle due to complex interdependencies.