Cleanup the gitlint configuration file
I decided to remove the examples so that we get a minimal file with only
the necessary/relevant information. This should be a refactoring (i.e.,
no functional changes).

I replaced the examples with a link to the documentation (which also
contains an up-to-date version of the sample .gitlint file), removed
ignore-merge-commits=true as this is the default, switched from IDs to
names for contrib rules (because names are more readable), and dropped
the unnecessary contrib-body-requires-signed-off-by section at the end.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>