actions/cargo: Make the matrix config for clippy a bit more readable
The old code for the matrix strategy was quite weird. A different syntax
was used for the `rust` and `optional` matrix variables and both lists
had only a single entry (which kinda defies the purpose).

It should be much more readable to explicitly specify all configurations
via `include` and such an example is also in the documentation [0].

[0]: https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs#example-adding-configurations

Signed-off-by: Michael Weiss <michael.weiss@atos.net>