actions/cargo: Allow unknown lints for the clippy check with the MSRV
We run clippy checks with the MSRV and beta toolchain. Only the check
with the MSRV is required as the other check could break due to an
update of the beta toolchain in the meantime (and it certainly shouldn't
block merging due to changes that are unrelated to the PR).

We also want to fix the job against the beta toolchain though as soon as
we notice new failures, which is regularly the case due to the addition
of new lints. Sometimes we may choose to ignore some of those newly
added lints (which is ok to do) but that would break the MSRV check as
"-D warnings" turns the "unknown-lints" warning into an error.
Therefore, we simply allow unknown lints for the MSRV check (but not for
the beta toolchain so that we notice if a lint would get removed, etc.).

This approach should be fine as it's still guaranteed that butido
compiles with the MSRV (it does at most cause a warning when running
clippy) and we use newer Rust versions for the development anyway (and
it's probably safe to assume that most developers use (at least) the
most recent stable version).

Signed-off-by: Michael Weiss <michael.weiss@atos.net>